--- index/region.split.index.test
@@ -1,10 +1,0 @@
-1	1	100000
-1	100001	200000
-1	200001	300000
-1	300001	400000
-1	400001	500000
-1	500001	600000
-1	600001	700000
-1	700001	800000
-1	800001	900000
-1	900001	1000000
